Change default s:back to s:base02
I have too many terminal tools that use bright black, i.e. base03,
which is too damn dark in vanilla solarized, that I've had to make
that brighter (val 100 or so), which in turn is too light for vim
background to my liking. Thus, the base02 make more sense to be
the vim background, which I've in turn made a bit darker
(val 42 or so).